Address requests changed in pull request
- Remove "action" since currently it is basically unused
- Move tests of FileOpener to tests/lib/test_filesystem
- Use path.write_text to initialize file
- Avoid comment in changes to requirements
