+++ src/pip/_internal/utils/temp_dir.py
@@ -125,16 +125,7 @@ class TempDirectory(object):
         """
         self._deleted = True
         if os.path.exists(self._path):
-            try:
-                rmtree(self._path)
-            except OSError as e:
-                if sys.platform == 'win32' and e.errno == errno.EACCES:
-                    logger.warning(
-                        "%s (virus scanner may be holding it)."
-                        "cannot remove '%s'",
-                        e.strerror, e.filename)
-                else:
-                    raise
+            rmtree(self._path)
 
 
 class AdjacentTempDirectory(TempDirectory):
